backend/requirements.txt
# FleetFix Backend Requirements

# Core Framework
fastapi==0.118.0
uvicorn[standard]==0.37.0
pydantic==2.11.9
pydantic-settings==2.11.0

# Database
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
alembic==1.16.5

# Data Processing
pandas==2.3.3
numpy==2.3.3

# Data Generation (for seed_data.py)
Faker==37.8.0

# Visualization
plotly==6.3.1

# LLM Integration
openai==2.1.0
anthropic==0.69.0

# HTTP & API
httpx==0.28.1
python-multipart==0.0.20

# Utilities
python-dotenv==1.1.1
python-dateutil==2.9.0.post0

# Development & Testing
pytest==8.4.2
pytest-asyncio==1.2.0
black==25.9.0
flake8==7.3.0